Ignore case (#8)
* Ignore case
